+++ src/Squidex.Domain.Apps.Events/Schemas/SchemaCreatedNestedField.cs
@@ -9,16 +9,7 @@ using Squidex.Domain.Apps.Core.Schemas;
 
 namespace Squidex.Domain.Apps.Events.Schemas
 {
-    public sealed class SchemaCreatedNestedField
+    public sealed class SchemaCreatedNestedField : SchemaCreatedFieldBase
     {
-        public string Name { get; set; }
-
-        public bool IsHidden { get; set; }
-
-        public bool IsLocked { get; set; }
-
-        public bool IsDisabled { get; set; }
-
-        public FieldProperties Properties { get; set; }
     }
 }